Class Make.base_server

The server baseclass

method virtual on_notification : notify_back: (Lsp.Server_notification.t -> unit IO.t) -> server_request:send_request -> Lsp.Client_notification.t -> unit IO.t

method virtual on_request : 'a. notify_back: (Lsp.Server_notification.t -> unit IO.t) -> server_request:send_request -> id:Req_id.t -> 'a Lsp.Client_request.t -> 'a IO.t

Method called to handle client requests.

parameter notify_back
an object used to reply to the client, send progress messages, diagnostics, etc.

parameter id
the query RPC ID, can be used for tracing, cancellation, etc.

method must_quit : bool

Set to true if the client requested to exit

method virtual spawn_query_handler : (unit -> unit IO.t) -> unit

How to start a new future/task/thread concurrently. This is used to process incoming user queries.