Module Cli.Channel

The channel the server shold use to listen for connections

type t =

| Stdio

| Pipe of string
(*
A path to the unix domain socket or windows pipe
*)

| Socket of int
(*
A tcp connection on localhost with the port number
*)