Package logs 0.10.0

Logs provides a logging infrastructure.

Logging is performed on sources whose reporting level can be set independently. Log message report is decoupled from logging and is handled by a reporter. A few optional log reporters are distributed with the package and the API easily allows to implement your own.

See the basics.

Library logs

Logs Logging.

Libraries logs.{browser,cli,fmt,lwt,threaded}

Each of these modules lives in its own library.

Logs_browser

Logs_cli Cmdliner support for Logs.

Logs_fmt Format colorful reporter for Logs.

Logs_lwt Lwt logging.

Logs_threaded Thread safe logging.

Package info

authors

The logs programmers

changes-files

CHANGES.md

depends

mtime

ocaml

ocamlbuild

ocamlfind

topkg

homepage

https://erratique.ch/software/logs

issues

https://github.com/dbuenzli/logs/issues

license

ISC

license-files

LICENSE.md

maintainers

Daniel Bünzli <daniel.buenzl i@erratique.ch>

online-doc

https://erratique.ch/software/logs/doc

readme-files

README.md

repo

git+https://erratique.ch/repos/logs.git

tags

log

org:erratique

system

version

0.10.0