Package uutf 1.0.4

Warning. You are encouraged not to use this library.

As of OCaml 4.14, both UTF encoding and decoding are available in the standard library, see the String and Buffer modules.

If you are looking for a stream abstraction compatible with effect based concurrency look into the bytesrw package.

Uutf is a non-blocking streaming codec to decode and encode the UTF-8, UTF-16, UTF-16LE and UTF-16BE encoding schemes. It can efficiently work character by character without blocking on IO. Decoders perform character position tracking and support newline normalization.

Functions are also provided to fold over the characters of UTF encoded OCaml string values and to directly encode characters in OCaml Buffer.t values.

Library uutf

Uutf Non-blocking streaming Unicode codec.

Package info

authors

The uutf programmers

changes-files

CHANGES.md

depends

ocaml

ocamlbuild

ocamlfind

topkg

homepage

https://erratique.ch/software/uutf

issues

https://github.com/dbuenzli/uutf/issues

license

ISC

license-files

LICENSE.md

maintainers

Daniel Bünzli <daniel.buenzl i@erratique.ch>

online-doc

https://erratique.ch/software/uutf/doc/

readme-files

README.md

repo

git+https://erratique.ch/repos/uutf.git

tags

codec

org:erratique

text

unicode

utf-16

utf-8

version

1.0.4