Package yojson 2.2.2

The Yojson library provides runtime functions for reading and writing JSON data from OCaml. The design goals of Yojson are the following:

Allowing type-aware serializers/deserializers to read and write directly without going through a generic JSON tree, for efficiency purposes.

Distinguishing between ints and floats.

Providing optional extensions of the JSON syntax. These extensions include comments, arbitrary strings, optional quotes around field names, tuples and variants.

See JSON specification.

Author: Martin Jambon

Entry point

The entry point of this library is the module Yojson.

Package info

authors

Martin Jambon

changes-files

CHANGES.md

depends

alcotest

dune

ocaml

odoc

seq

homepage

https://github.com/ocaml-community/yojson

issues

https://github.com/ocaml-community/yojson/issues

license

BSD-3-Clause

license-files

LICENSE.md

maintainers

marek@tarides.com

nathan@tarides.com

paul-elliot@tarides.com

online-doc

https://ocaml.org/p/yojson/latest

readme-files

README.md

repo

git+https://github.com/ocaml-community/yojson.git

version

2.2.2